🔄 Rename tfcoil module to tf_coil for consistency across the codebase
